Use fqdn instead of IP address to connect to vault
Fixes #69 by using the FQDN to connect to the vault instance instead of
using the IP address.

Using the IP address caused an issue because the IP address would change
after a crash or removal of a pod. In turn, this would cause the TLS
certificate to no longer be valid, as the TLS certificate is validated
against the new IP address (but was only issued for the old IP address).
We could re-issue the certificate, but the certificate is also valid for
anything which uses the same FQDN which the new pod will share.

This change removes the IP address from the certificate, and relies on
the FQDN instead.