Decrease the vault_client charm library version
Due to an intermittent issue with the integration tests, version 13 was never published, so now 14 will not publish as the version numbers must be consecutive. I've tried re-building version 13 several times but so far have not had any luck getting the tests to succeed, so this seems like the path of least resistance.