Removes unittest, so pytest operates as expected
Currently, we have a mashup of unittest and pytest.

Although the class inherits from unittest, we're using pytest to run
the tests. This works fine in the basic cases, but doesn't work when you
try to use any of the features of either framework. For example, neither
pytest parameterize or unitttest subtests work.